Demo/service-examples/Carta-People-Counting/common/requirements.txt
tensorflow
numpy
pandas
sklearn
matplotlib
h5py
keras
pulsar-client~=2.7.2
opencv-python-headless